Refactored the code by moving the database query method calls to the ParseClientData class, which now retrieves raw data from the database and return it to the SendDataToClient class already in QByteArray ready-to-send
This change allows the sending class to focus solely on transmitting data without needing to verify it after receiving user input, as the parsing class handles all checks and sends ready-to-send data in the QByteArray format.

Additionally, this refactor enables the SQL query class to simply query data and return raw results without additional conversion to QJsonDocument.

As a result, the RequestHandler file, which was solely responsible for converting database data to compact JSON format, has been removed.